move task action to separate file
